Refactor: Move UserInfo classes to own file.
